requirements.txt
pandas==2.2.2
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2
streamlit==1.36.0
numpy==1.26.4
lightgbm==4.3.0
scikit-surprise==1.1.4